feat(shortcodes/card): make links smarter and fix styling
The card link was overlaying the entire page because the link element
has position absolute with no close parent having a position relative.

Also, removed the box-shadow from the link to not have a line at the
bottom of the card element.

Finally, links are now smarter, so if a relative link is added, it uses
the relref function, which will give an error at build time if the
referenced page does not exist.